+++ test/lib/resources/accounts-test.js
@@ -2,6 +2,7 @@
 
 const expect = require('expect.js');
 const client = require('../../../index')('sk-XXX').accounts;
+const constants = require('../../../lib/constants');
 let nock = require('nock');
 
 describe('lib/resources/accounts', function() {
@@ -53,13 +54,14 @@ describe('lib/resources/accounts', function() {
 
     it('accepts an options object', function() {
       this.scope = nock('https://api.usebutton.com:443')
-        .get('/v1/affiliation/accounts/' + this.accountId + '/transactions?cursor=cursor&start=2015-01-01T00%3A00%3A00Z&end=2016-01-01T00%3A00%3A00Z')
+        .get('/v1/affiliation/accounts/' + this.accountId + '/transactions?cursor=cursor&start=2015-01-01T00%3A00%3A00Z&end=2016-01-01T00%3A00%3A00Z&time_field=modified_date')
         .reply(200, { meta: { status: 'ok' }, 'objects': this.transactions });
 
       return client.transactions(this.accountId, {
         cursor: 'cursor',
         start: '2015-01-01T00:00:00Z',
-        end: '2016-01-01T00:00:00Z'
+        end: '2016-01-01T00:00:00Z',
+        time_field: constants.TIME_FIELD_MODIFIED
       }).then((result) => {
         expect(result.data).to.eql(this.transactions);
         this.scope.done();